fix: fix camo visibility, improve obstacle prone visibility

+++ src/monster.cpp
@@ -1451,7 +1451,7 @@ bool monster::sees( const Creature &ch ) const
     }
     return Creature::sees( ch );
 }
-bool monster::sees( const tripoint_bub_ms &t, bool is_player, int range_mod ) const
+bool monster::sees( const tripoint_bub_ms &t, bool is_player, double range_mod ) const
 {
     if( type->clairvoyance > 0 ) {
         const int wanted_range = rl_dist( bub_pos(), t );
